# improve rolling_window performance