This slightly improves rolling_window functions, `rolling_min, rolling_max, rolling_mean` by removing an expensive modulo in a hot loop and using trusted_length iter collect.

before
```
bench rolling rate: 0%  time:   [134.85 us 135.17 us 135.47 us]  
```
after
```
bench rolling rate: 0%  time:   [104.08 us 104.68 us 105.49 us] 
```